+++ src/Sentry.Compiler.Extensions/DiagnosticCategories.cs
@@ -1,0 +1,6 @@
+namespace Sentry.Compiler.Extensions;
+
+internal static class DiagnosticCategories
+{
+    internal const string Sentry = nameof(Sentry);
+}